--- a/app/models/diary.rb
+++ b/app/models/diary.rb
@@ -8,6 +8,9 @@ class Diary < ApplicationRecord
   scope :public_diaries, -> { where(is_public: true) }
   scope :private_diaries, -> { where(is_public: false) }
 
+  # 統計チャートのキャッシュを無効化
+  after_commit :clear_stats_cache
+
   def github_uploaded?
     github_uploaded == true
   end
@@ -20,5 +23,11 @@ class Diary < ApplicationRecord
     return nil unless selected_til_index.present?
 
     til_candidates.find_by(index: selected_til_index)&.content
+  end
+
+  private
+
+  def clear_stats_cache
+    Rails.cache.delete_matched("stats_charts_#{user_id}_*")
   end
 end