… Gemfile.lock
… app/views/layouts/application.html.erb
… config/credentials.yml.enc
… config/database.yml
… config/environments/development.rb
… config/environments/production.rb
… config/environments/test.rb
… config/puma.rb
… spec/models/question_cache_simple_spec.rb
… spec/models/question_cache_spec.rb
… spec/requests/stats_cache_simple_spec.rb
… spec/requests/stats_cache_spec.rb